% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayLabelArray.R
\name{DisplayLabArr}
\alias{DisplayLabArr}
\title{Plot 3d label array}
\usage{
DisplayLabArr(label.array, sl.dim = 2, foldername = "Images",
  filename_pre = "Image", bg.index = NA, bg.col = "black",
  border.col = NA)
}
\arguments{
\item{label.array}{Label array to be displayed.}

\item{sl.dim}{The dimension that will be used for slicing. The default is 2nd dimension.}

\item{foldername}{Name of the folder that will contain the result images.}

\item{filename_pre}{The prefix for the filename of result images. A result image will 
have name \emph{filename_pre.n.jpg} where \emph{n} is the slice index.}

\item{bg.index}{The logical matrix of the same size as \emph{label.array} specifying the 
background. Background cells have value TRUE and non-background cells have
value FALSE. All the background cells will be display with one common color specified
by \emph{bg.col}.}

\item{bg.col}{Color for the background. The default is black.}

\item{border.col}{Color for cell borders. If NA, not showing cell border.}
}
\value{
A new folder is created in working directory, containing slice-by-slice images of 
the 3D label array.
}
\description{
Display 3d label array slice by slice, different labels will have different colors.
}
\examples{
data(LabelArrayExample)
bg.index <- array(LabelArrayExample==0,dim(LabelArrayExample))
DisplayLabArr(label.array=LabelArrayExample, bg.index = bg.index, 
foldername="Images", filename_pre="Image")
}

