% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dla_sim.R
\name{dla}
\alias{dla}
\title{Generate DLA cluster}
\usage{
dla(N = 5000, initial = data.frame(x = 0, y = 0), source = Inf,
  abandon.factor = 10, diameter.tol = 0.01, checkin.every = 5000,
  verbose = TRUE, write.to = NULL, phi = c(0, 2 * pi), theta = c(0, pi))
}
\arguments{
\item{N}{number of particles to be added to the initial cluster (can be Inf)}

\item{initial}{a data-frame or matrix of coordinates of the initial cluster. For a single seed particle, a coordinate vector will be also accepted. The default is a single seed particle at the origin of 2-dimensional plane.
The number of columns must be 2 or 3 and serves as the dimensionality of space in which the cluster will be grown.
Column names do not matter, only their order.}

\item{source}{a vector of coordinates of the particle source. A single value Inf is understood, but apart from that, must be the same length as the number of columns in \code{initial}.}

\item{abandon.factor}{a multiplying coefficient controlling when straying particles should be abandoned. The max cluster distance from the origin or the finite source distance from the origin, whichever is greater, are multiplied by this factor to get the abandoning distance.
If a particle strays farther than the abandoning distance from the cluster, it is abandoned.}

\item{diameter.tol}{a fraction of the particle diameter representing the sticking penetration tolerance.}

\item{checkin.every}{if not NULL or Inf, growing the cluster will be split into stages of adding this many particles.
At the end of each stage the function will check whether the source has been blocked by the cluster (if yes - will message stop).
It may also print a progress message and export cluster to a file, as controlled by the arguments \code{verbose} and \code{write.to}.}

\item{verbose}{logical, indicating whether the function should print out progress statements at the end of each stage controlled by \code{checkin.every}.}

\item{write.to}{if not NULL, it is the file path to which the function will export the cluster at the end of each stage controlled by \code{checkin.every}.}

\item{phi, theta}{two vectors of length 2, denoting the boundaries of the sector of particle propagation. \code{phi} is the azimuthal angle (longitude in 3D) and \code{theta} is the polar angle of spherical coordinates in 3D (the pole being 0).}
}
\value{
A dataframe containing coordinates of cluster particles. The number of rows will be the number of points in the initial cluster plus \code{N}.
}
\description{
Diffusion-limited aggregation (DLA) cluster can be generated in 2- or 3-dimensional Euclidean space.
}
