% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dla_sim.R
\name{dla2d}
\alias{dla2d}
\title{2D Diffusion-limited Aggregation}
\usage{
dla2d(N = 200, initial = data.frame(x = 0, y = 0), abandon.factor = 10,
  diameter.tol = 0.01, write.every = 3600, write.to = NULL)
}
\arguments{
\item{N}{number of particles to be added to the pre-existing cluster (could be Inf)}

\item{initial}{a data-frame or matrix with two columns, containing coordinates of a pre-existing cluster.}

\item{abandon.factor}{a multiplier of the current containing radius of the cluster to get the distance beyond which particles are abandoned}

\item{diameter.tol}{tolerance fraction of the particle diameter in sticking (penetration)}

\item{write.every}{a period of regular exporting the current cluster to a file, in addition to writing it last time upon reaching N particles}

\item{write.to}{path of the file to export the current cluster to. If NULL, there will be no exporting of data.}
}
\value{
a data-frame of cluster coordinates, same as finally written to \code{write.to}
}
\description{
2D Diffusion-limited Aggregation
}
\details{
Particle diameter is 1. The default value of \code{initial} means that the pre-existing cluster is a single seed particle at the origin.
}
