% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LabelArrayExample}
\alias{LabelArrayExample}
\title{3D Label Array for Magnetic Resonance (MR) Brain Images}
\format{A 86x11x91 array of integers. Background voxels are labelled by value 0.}
\usage{
data(LabelArrayExample)
}
\description{
The data set is a three-dimension array. This is the label array resulting from segmenting the average of 3D 
MR brain images of 300 people with Alzheimer's disease. The segmentation method used
is Segmentation by Weighted Aggregation (SWA)
\url{http://www.nature.com/nature/journal/v442/n7104/full/nature04977.html}
}
\keyword{datasets}

