% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sle.R
\name{sle}
\alias{sle}
\title{Stochastic Loewner Evolution}
\usage{
sle(kappa = 4, tmax = 1, a = 1, kappaL = 0, nsteps = 2000,
  p_timescaling = 0.5, verbose = TRUE, forcing = NULL)
}
\arguments{
\item{kappa}{strength of the 1d Brownian motion in the driving function.}

\item{tmax}{max time to which the trace is developed.}

\item{a}{exponent of the Levy flights component in the driving function.}

\item{kappaL}{strength of the Levy flights component in the driving function.}

\item{nsteps}{number of steps.}

\item{p_timescaling}{exponent determining the distribution of time steps. The succession of times will be made uniform in the variable \code{t^p_timescaling}.}

\item{verbose}{boolean, to print progress statements or not.}

\item{forcing}{if not NULL, should be a dataframe of the driving function, which will then be used, overriding other driving-related arguments.
The dataframe should have columns t and xi, starting from t=0 and xi=0, sorted.}
}
\value{
List with components: \code{t} - vector of time values, \code{xi} - vector of values of the driving function, \code{gamma} - data frame of x and y coordinates of the generated trace,\code{t_cross} - crossover time between Brownian and Levy components (NULL if \code{kappaL = 0}), \code{call_params} - list of call parameters, \code{runtime} - elapsed time in seconds.
}
\description{
Generate SLE trace driven by Brownian motion, possibly with the addition of Levy-flights
}
\note{
SLE (Stochastic Loewner Evolution) is a generative stochastic process for growing a stocastic curve (trace) out of a boundary of a 2D domain.
It uses a continuous family of conformal maps, parametrized by a "time" parameter: w(z,t). The SLE equation is dw(z,t)/dt = 2/(w(z,t)-xi(t)), w(z,0)=z. Here xi(t) is the real-valued driving function of the process, assumed to be a stochastic process.
The mapping w(z,t) is from the upper half plane. In the standard SLE, xi(t) is the 1D Brownian motion with diffusion constant kappa (and an intricate connection to conformal field theory exists).
As a generalization, we also allow xi(t) to be a sum of 1D Brownian motion and Levy fligts.

For more details see this publication and references therein:

Rushkin, I., Oikonomou, P., Kadanoff, L.P. and Gruzberg, I.A., 2006. Stochastic Loewner evolution driven by Lévy processes. Journal of Statistical Mechanics: Theory and Experiment, 2006(01), p.P01001.
}
