% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dla_sim.R
\name{vis.dla}
\alias{vis.dla}
\title{Visualize DLA cluster}
\usage{
vis.dla(data, color0 = "green", color1 = "red")
}
\arguments{
\item{data}{a dataframe or matrix with 2 or 3 columns of particle coordinates (for a 2D or 3D cluster), like the one produced by \code{dla2d} or \code{dla3d}.}

\item{color0}{color for the earliest particles}

\item{color1}{color for the latest particles}
}
\description{
Visualize DLA cluster
}
