% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dla_sim.R
\name{dla}
\alias{dla}
\title{Generate DLA cluster}
\usage{
dla(N = 200, initial = data.frame(x = 0, y = 0), source = Inf,
  abandon.factor = 10, diameter.tol = 0.01, report.every = NULL,
  verbose = TRUE, write.to = NULL)
}
\arguments{
\item{N}{number of particles to be added to the initial cluster (can be Inf)}

\item{initial}{a data-frame or matrix of coordinates of the initial cluster. The default is a single seed particle at the origin of 2-dimensional plane.
The number of columns must be 2 or 3 and serves as the dimensionality of space in which the cluster will be grown.
Column names do not matter, only their order.}

\item{source}{a vector of coordinates of the particle source. A single value Inf is understood, but apart from that, must be the same length as the number of columns in \code{initial}.}

\item{abandon.factor}{a multiplying coefficient controlling when straying particles should be abandoned. The max cluster distance from the origin or the finite source distance from the origin, whichever is greater, are multiplied by this factor to get the abandoning distance.
If a particle strays farther than the abandoning distance from the cluster, it is abandoned.}

\item{diameter.tol}{a fraction of the particle diameter representing the sticking penetration tolerance.}

\item{report.every}{if not NULL or Inf, growing the cluster will be split into stages of adding this many particles.}

\item{verbose}{logical, indicating whether the function should print out progress statements at the end of each stage controlled by \code{report.every}.}

\item{write.to}{if not NULL, it is the file path to which the function will save the cluster at the end of each stage controlled by \code{report.every}.}
}
\value{
A dataframe containing coordinates of cluster particles. The number of rows will be the number of points in the initial cluster plus \code{N}.
}
\description{
Diffusion-limited aggregation (DLA) cluster can be generated in 2- or 3-dimensional Euclidean space.
}
