% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dla_sim.R
\name{vis.dla}
\alias{vis.dla}
\title{Visualize DLA cluster}
\usage{
vis.dla(data, color0 = "green", color1 = "red", return.data = FALSE,
  plot = TRUE, dupl.rm = TRUE)
}
\arguments{
\item{data}{a dataframe or matrix with 2 or 3 columns of particle coordinates (for a 2D or 3D cluster), like the one produced by \code{dla2d} or \code{dla3d}. An additional column named "color" will be tolerated and ignored.}

\item{color0}{color for the earliest particles, start color of the color scale}

\item{color1}{color for the latest particles, end color of the color scale}

\item{return.data}{logical, whether the function should return the data frame used for plotting}

\item{plot}{logical, whether the function should produce a plot}

\item{dupl.rm}{logical, whether to remove duplicate particles (particles with identical coordinates), leaving the earliest ones. Normally, there should not be duplicate particles, but one exception is when the source of particles was at a finite point and the cluster reached it, so all particles emitted after that remain at the source.}
}
\value{
if \code{return.data} is TRUE, will return the data frame with the extra column "color" containing particle colors
}
\description{
Particles will be colored in the chronological order using a color scale. The function can  produce a plot and/or return the data frame of the cluster coordinates with the particle colors added.
}
